+++ jale.sh
@@ -1,7 +1,7 @@
 #!/bin/bash
 
 # Just Another Linux Enumerator
-version="version 0.0.2"
+version="version 0.0.3"
 
 # run as low-privilege user
 # inspired by LinEnum
@@ -9,7 +9,6 @@ version="version 0.0.2"
 # usage and help
 
 usage() {
-    echo
     echo "++++++++++++++++++++++++++++++++++++++++"
     echo "| Just Another Linux Enumerator        |"
     echo "|                github: kegn          |"
@@ -23,17 +22,34 @@ usage() {
     echo "++++++++++++++++++++++++++++++++++++++++"
 }
 
+# color definitions
+green() {
+    echo -e "\e[32m$*"
+}
+
+yellow() {
+    echo -e "\e[33m$*"
+}
+
+cyan() {
+    echo -e "\e[36m$*"
+}
+
+# reset echo foreground color
+echofg() {
+    echo -e "\e[39m$*"
+}
 header() {
-     echo "++++++++++++++++++++++++++++++++++++++++"
-     echo "| Just Another Linux Enumerator        |"
-     echo "++++++++++++++++++++++++++++++++++++++++"
+     green "++++++++++++++++++++++++++++++++++++++++" 2>/dev/null
+     yellow "  Just Another Linux Enumerator         " 2>/dev/null
+     green "++++++++++++++++++++++++++++++++++++++++" 2>/dev/null
 }
 
 target_user() {
-         echo "\n\n"
-         echo "++++++++++++++++++++++++++++++++++++++++"
-         echo "| Target User: $TARGET_USER            |"
-         echo "++++++++++++++++++++++++++++++++++++++++"
+echo ""        
+green "++++++++++++++++++++++++++++++++++++++++" 2>/dev/null
+yellow " Target User: $TARGET_USER             " 2>/dev/null
+green "++++++++++++++++++++++++++++++++++++++++" 2>/dev/null
 }
 
 # for proper output spacing
@@ -42,8 +58,9 @@ sed -e 's/^/               /g' 2>/dev/null
 }
 
 system_info() {
-echo "\n"
-echo " ----- System Information -----"
+echo ""
+cyan " ----- System Information -----" 2>/dev/null
+echofg 2>/dev/null
 echo "hostname     : $(cat /etc/hostname 2>/dev/null)"
 echo "current user : $(whoami 2>/dev/null)"
 echo "uname -a     : $(uname -a 2>/dev/null)"
@@ -53,11 +70,12 @@ echo "shells       : "
 cat /etc/shells | grep -v "#" 2>/dev/null | spacing
 echo "release info : "
 echo "$(cat /etc/*-release 2>/dev/null | spacing)"
-echo ""
 }
 
 user_info() {
-echo " ----- User Information ----- "
+echo ""
+cyan " ----- User Information ----- " 2>/dev/null
+echofg 2>/dev/null
 echo "current user : $(whoami 2>/dev/null)"
 echo "id           : $(id 2>/dev/null)"
 echo "PATH         : $(echo $PATH 2>/dev/null)"
@@ -72,7 +90,9 @@ echo "$(w 2>/dev/null | spacing)"
 }
 
 common_files() {
-echo " ----- Common Files ----- "
+echo ""
+cyan " ----- Common Files ----- " 2>/dev/null
+echofg 2>/dev/null
 echo "/etc/passwd: "
 cat /etc/passwd 2>/dev/null | spacing
 echo ""
@@ -80,6 +100,7 @@ echo ""
 shadow=$(cat /etc/shadow 2>/dev/null)
 if [ "$shadow" ];
 then
+    echo ""
     echo "/etc/shadow: "
     cat /etc/shadow 2>/dev/null | spacing
 fi
@@ -87,6 +108,7 @@ fi
 sudoers=$(cat /etc/sudoers 2>/dev/null)
 if [ "$sudoers" ];
 then
+    echo ""
     echo "/etc/sudoers: "
     cat /etc/sudoers 2>/dev/null | grep -v "#" 2>/dev/null | spacing
 fi
@@ -94,7 +116,9 @@ fi
 
 # fix to / instead of .
 writable_files() {
-echo " ----- Writable Files ----- "
+echo ""
+cyan " ----- Writable Files ----- " 2>/dev/null
+echofg 2>/dev/null
 echo "files that we do NOT own:"
 find / -writable ! -user $(whoami) -type f ! -path "proc/*" ! -path "/sys/*" -exec ls -alh {} \; 2>/dev/null | spacing
 echo ""
@@ -105,7 +129,9 @@ find / -writable -user $(whoami) -type f ! -path "proc/*" ! -path "/sys/*" -exec
 
 # fix to / instead of .
 writable_dirs() {
-echo " ----- Writable Directories ----- "
+echo ""
+cyan echo " ----- Writable Directories ----- " 2>/dev/null
+echofg 2>/dev/null
 echo "directories that we do NOT own:"
 find / -writable ! -user $(whoami) -type d ! -path "proc/*" ! -path "/sys/*" 2>/dev/null | spacing
 echo ""
@@ -114,18 +140,24 @@ find / -writable -user $(whoami) -type d ! -path "proc/*" ! -path "/sys/*" 2>/de
 }
 
 hidden_files() {
-echo " ----- Hidden Files ----- "
+echo ""
+cyan " ----- Hidden Files ----- " 2>/dev/null
+echofg 2>/dev/null
 echo "hidden files:"
 find / -name ".*" -type f ! -path "proc/*" ! -path "/sys/*" -exec ls -alh {} \; 2>/dev/null | spacing
 }
 
 environment() {
-echo " ----- Environment ----- "
+echo ""
+cyan " ----- Environment ----- "
+echofg 2>/dev/null
 env 2>/dev/null | grep -v 'LS_COLORS' 2>/dev/null | spacing
 }
 
 network_info() {
-echo " ----- Network Info -----"
+echo ""
+cyan " ----- Network Info -----" 2>/dev/null
+echofg 2>/dev/null
 echo "ip info:"
 /sbin/ifconfig -a 2>/dev/null | spacing
 echo "arp info:"
@@ -142,6 +174,9 @@ netstat -ulpn 2>/dev/null
 }
 
 svc_info() {
+echo ""
+cyan " ----- Service Info ----- " 2>/dev/null
+echofg 2>/dev/null
 echo "processes: ps aux: "
 ps aux 2>/dev/null
 
@@ -151,21 +186,29 @@ ps aux 2>/dev/null | awk '{print $11}' | xargs -r ls -la 2>/dev/null | awk '!x[$
 }
 
 software_versions() {
-    echo "MYSQL Version: $(mysql --version 2>/dev/null)"
+echo ""
+cyan " ----- Software Versions ----- " 2>/dev/null
+echofg 2>/dev/null
+echo "MYSQL Version    : $(mysql --version 2>/dev/null)"
+echo "Apache Version   : $(apache2 -v 2>/dev/null | head -n 1 2>/dev/null)"
+echo "Docker Version   : $(docker --version 2>/dev/null)"
+echo "Sudo Version     : $(sudo -V 2>/dev/null | grep 'Sudo version')"
     
 }
 
-enabled_websites() {
-    # check apache2 default config
-    ls /etc/apache2 > /dev/null 2>&1
-    if [ $? -eq 0 ]
-    then
-        echo "apache2 site enabled:"
-        cat /etc/apache2/sites-enabled/* | grep "Document"
 
-    fi
-
-}
+# eventually add discovery of virtualhosts
+# enabled_websites() {
+#     # check apache2 default config
+#     ls /etc/apache2 > /dev/null 2>&1
+#     if [ $? -eq 0 ]
+#     then
+#         echo "apache2 site enabled:"
+#         cat /etc/apache2/sites-enabled/* | grep "Document"
+# 
+#     fi
+# 
+# }
 
 
 #### MAIN SCRIPT ####