+++ CHANGELOG.md
@@ -4,3 +4,7 @@
   - inital version
 ## [v0.0.2] - 2020-05-21
   - fix parameter specification in usage()
+## [v0.0.3] - 2020-05-22
+  - add experimental colors cyan, yellow, green, default
+  - colorize section headers
+  - add checks for more software versions